more consistent types, some bounds checking

+++ compress.h
@@ -21,10 +21,10 @@ extern "C" {
 #define RUN_SIZE      32
 #define LONG_RUN_SIZE 1024
 
-size_t pack   (uint8_t *unpacked, uint32_t inputsize, uint8_t *packed, int fast);
+size_t pack   (uint8_t *unpacked, size_t inputsize, uint8_t *packed, int fast);
 size_t unpack (uint8_t *packed, uint8_t *unpacked);
 
-size_t unpack_from_file (FILE *file, unsigned int offset, uint8_t *unpacked);
+size_t unpack_from_file (FILE *file, size_t offset, uint8_t *unpacked);
 
 #ifdef __cplusplus
 }
